pyproject.toml
[project]
name = "mcp-snowflake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastmcp>=2.9.0",
  "snowflake-connector-python>=3.15.0",
  "snowflake-core>=1.6.0",
]
